+++ doc/source/code/tut3.py
@@ -1,4 +1,4 @@
-from netpyne import init
+from netpyne import sim
 
 # Network parameters
 netParams = {}  # dictionary to store sets of network parameters
@@ -35,7 +35,7 @@ netParams['connParams'].append({'preTags': {'popLabel': 'S'}, 'postTags': {'popL
 	'weight': 0.01, 			# synaptic weight 
 	'delay': 5,					# transmission delay (ms) 
 	'sec': 'dend',				# section to connect to
-	'loc': '1.0',				# location of synapse
+	'loc': 1.0,				# location of synapse
 	'synMech': 'exc'})   		# target synaptic mechanism
 netParams['connParams'].append({'preTags': {'popLabel': 'background'}, 'postTags': {'cellType': 'PYR'}, # background -> PYR
 	'weight': 0.01, 				# synaptic weight 
@@ -58,6 +58,6 @@ simConfig['plot2Dnet'] = True           # plot 2D visualization of cell position
 
 
 # Create network and run simulation
-init.createAndSimulate(netParams = netParams, simConfig = simConfig)    
+sim.createAndSimulate(netParams = netParams, simConfig = simConfig)    
    
 # import pylab; pylab.show()  # this line is only necessary in certain systems where figures appear empty